requirements.txt
Flask
requests
xmltodict
python-dotenv
pynetbox